make fmi-base modeldescription only. common package moved to importer
read and write on variables are now handled by extension functions
