+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/common/FmuInstance.kt
@@ -26,6 +26,8 @@ package no.ntnu.ihb.fmi4j.common
 
 import no.ntnu.ihb.fmi4j.modeldescription.CommonModelDescription
 import no.ntnu.ihb.fmi4j.modeldescription.ModelDescription
+import no.ntnu.ihb.fmi4j.modeldescription.RealArray
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReferences
 import no.ntnu.ihb.fmi4j.modeldescription.variables.ModelVariables
 import java.io.Closeable
