+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/cs/CoSimulationLibraryWrapper.kt
@@ -25,7 +25,7 @@
 package no.ntnu.ihb.fmi4j.importer.cs
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus
-import no.ntnu.ihb.fmi4j.common.ValueReferences
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReferences
 import no.ntnu.ihb.fmi4j.importer.jni.*
 
 /**
