+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/cs/CoSimulationSlave.kt
@@ -26,7 +26,7 @@ package no.ntnu.ihb.fmi4j.importer.cs
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus
 import no.ntnu.ihb.fmi4j.common.FmuSlave
-import no.ntnu.ihb.fmi4j.common.ValueReferences
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReferences
 import no.ntnu.ihb.fmi4j.importer.AbstractFmuInstance
 import no.ntnu.ihb.fmi4j.modeldescription.CoSimulationModelDescription
 import org.slf4j.Logger
