+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/jni/Fmi2Library.kt
@@ -25,7 +25,7 @@
 package no.ntnu.ihb.fmi4j.importer.jni
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus
-import no.ntnu.ihb.fmi4j.common.ValueReferences
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReferences
 import no.ntnu.ihb.fmi4j.util.OsUtil
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
