+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/jni/Fmi2LibraryWrapper.kt
@@ -26,6 +26,9 @@ package no.ntnu.ihb.fmi4j.importer.jni
 
 import no.ntnu.ihb.fmi4j.common.*
 import no.ntnu.ihb.fmi4j.importer.misc.ArrayBuffers
+import no.ntnu.ihb.fmi4j.modeldescription.StringArray
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReference
+import no.ntnu.ihb.fmi4j.modeldescription.ValueReferences
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
