+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/util/OsUtil.kt
