+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/cs/vendors/twentysim/ControlledTemperatureTest.kt
@@ -1,6 +1,7 @@
 package no.ntnu.ihb.fmi4j.importer.cs.vendors.twentysim
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus
+import no.ntnu.ihb.fmi4j.common.read
 import no.ntnu.ihb.fmi4j.importer.TestFMUs
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
