+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/me/vendors/openmodelica/FmuExportCrossCompile.kt
@@ -1,6 +1,7 @@
 package no.ntnu.ihb.fmi4j.importer.me.vendors.openmodelica
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus
+import no.ntnu.ihb.fmi4j.common.read
 import no.ntnu.ihb.fmi4j.importer.TestFMUs
 import no.ntnu.ihb.fmi4j.solvers.Solver
 import no.ntnu.ihb.fmi4j.solvers.apache.ApacheSolver
