+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/misc/Benchmark.kt
@@ -1,5 +1,6 @@
 package no.ntnu.ihb.fmi4j.importer.misc
 
+import no.ntnu.ihb.fmi4j.common.read
 import no.ntnu.ihb.fmi4j.importer.Fmu
 import no.ntnu.ihb.fmi4j.importer.TestFMUs
 import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable
